Rename `svgLoader` to `loader`
