+++ lib/loader.js
