+++ lib/main.js
@@ -1,4 +1,4 @@
-import svgLoader from "./svgLoader.js";
+import svgLoader from "./loader.js";
 import Icon from "./components/Icon.js";
 import Sprite from "./components/Sprite.js";
 import Theme from "./components/Theme.js";
